+++ src/vmware/azext_vmware/commands.py
@@ -20,11 +20,20 @@ def load_command_table(self, _):
         g.custom_command('create', 'privatecloud_create')
         g.custom_command('update', 'privatecloud_update')
         g.custom_command('delete', 'privatecloud_delete')
-        g.custom_command('listadmincredentials', 'privatecloud_listadmincredentials')
-        g.custom_command('addidentitysource', 'privatecloud_addidentitysource')
-        g.custom_command('deleteidentitysource', 'privatecloud_deleteidentitysource')
+        g.custom_command('list-admin-credentials', 'privatecloud_listadmincredentials')
+        g.custom_command('add-identity-source', 'privatecloud_addidentitysource')
+        g.custom_command('delete-identity-source', 'privatecloud_deleteidentitysource')
+        g.custom_command('add-availability-zone', 'privatecloud_addavailabilityzone')
+        g.custom_command('delete-availability-zone', 'privatecloud_deleteavailabilityzone')
+        g.custom_command('add-cmk-encryption', 'privatecloud_addcmkencryption')
+        g.custom_command('delete-cmk-encryption', 'privatecloud_deletecmkenryption')
         g.custom_command('rotate-vcenter-password', 'privatecloud_rotate_vcenter_password')
         g.custom_command('rotate-nsxt-password', 'privatecloud_rotate_nsxt_password')
+
+    with self.command_group('vmware private-cloud identity', vmware_sdk, client_factory=cf_vmware) as g:
+        g.custom_command('assign', 'privatecloud_identity_assign')
+        g.custom_command('remove', 'privatecloud_identity_remove')
+        g.custom_show_command('show', 'privatecloud_identity_get')
 
     with self.command_group('vmware cluster', vmware_sdk, client_factory=cf_vmware) as g:
         g.custom_command('create', 'cluster_create')