+++ src/vmware/azext_vmware/tests/latest/test_placement_policies_scenario.py
@@ -38,7 +38,7 @@ class VmwarePlacementPolicyScenarioTest(ScenarioTest):
         placementPolicyVmHostUpdate = self.cmd('az vmware placement-policy vm-host update --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name} --state {state} --vm-members {vm_members} --host-members {host_members}').get_output_in_json()
         self.assertEqual(placementPolicyVmHostUpdate['name'], 'policy1')
 
-        placementPolicyVmHostDelete = self.cmd('az vmware placement-policy vm-host delete --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name}').output
+        placementPolicyVmHostDelete = self.cmd('az vmware placement-policy vm-host delete --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name} --yes').output
         self.assertEqual(len(placementPolicyVmHostDelete), 0)
 
         placementPolicyVmCreate = self.cmd('az vmware placement-policy vm create --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name} --state {state} --display-name {display_name} --vm-members {vm_members} --affinity-type {affinity_type}').get_output_in_json()
@@ -47,5 +47,5 @@ class VmwarePlacementPolicyScenarioTest(ScenarioTest):
         placementPolicyVmUpdate = self.cmd('az vmware placement-policy vm update --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name} --state {state} --vm-members {vm_members}').get_output_in_json()
         self.assertEqual(placementPolicyVmUpdate['name'], 'policy1')
 
-        placementPolicyVmDelete = self.cmd('az vmware placement-policy vm delete --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name}').output
+        placementPolicyVmDelete = self.cmd('az vmware placement-policy vm delete --resource-group {rg} --private-cloud {privatecloud} --cluster-name {cluster_name} --placement-policy-name {placement_policy_name}  --yes').output
         self.assertEqual(len(placementPolicyVmDelete), 0)